requirements.txt
beautifulsoup4>=4.12.0
requests>=2.31.0
tqdm>=4.66.0
numpy>=1.24.0
pandas>=2.0.0
rank-bm25>=0.2.2
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
transformers>=4.43.0
accelerate>=0.30.0
torch>=2.1.0
pypdf>=4.0.0
pdfplumber>=0.11.0
PyYAML>=6.0.0

